Fix L12 and R7
Fix for L12 Redundant code and R7 Redundant "else".
There is a well-known pattern "if a: return one else: return two".
In fact, the else keyword is redundant, because when first if condition
evaluates to True, the function returns, otherwise control flow goes on.
Removing else makes the code cleaner and easier to read.